refactor: extract AppContainer
