+++ packages/TesterApp/src/App.tsx
@@ -1,10 +1,6 @@
 import React from 'react';
-import { SafeAreaView, ScrollView, StatusBar } from 'react-native';
 
-import {
-  useIsDarkMode,
-  getDefaultBackgroundStyle,
-} from './components/ui/utils';
+import { AppContainer } from './components/ui/AppContainer';
 import { Section } from './components/ui/Section';
 import { SectionContainer } from './components/ui/SectionContainer';
 
@@ -13,29 +9,20 @@ import { AsyncContainer } from './components/asyncChunks/AsyncContainer';
 import { MiniAppsContainer } from './components/miniapp/MiniAppsContainer';
 
 const App = () => {
-  const isDarkMode = useIsDarkMode();
-  const backgroundStyle = getDefaultBackgroundStyle(isDarkMode);
-
   return (
-    <SafeAreaView style={backgroundStyle}>
-      <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
-      <ScrollView
-        contentInsetAdjustmentBehavior="automatic"
-        style={backgroundStyle}
-      >
-        <SectionContainer>
-          <Section title="Async chunk">
-            <AsyncContainer />
-          </Section>
-          <Section title="Remote chunks">
-            <RemoteContainer />
-          </Section>
-          <Section title="Mini-apps">
-            <MiniAppsContainer />
-          </Section>
-        </SectionContainer>
-      </ScrollView>
-    </SafeAreaView>
+    <AppContainer>
+      <SectionContainer>
+        <Section title="Async chunk">
+          <AsyncContainer />
+        </Section>
+        <Section title="Remote chunks">
+          <RemoteContainer />
+        </Section>
+        <Section title="Mini-apps">
+          <MiniAppsContainer />
+        </Section>
+      </SectionContainer>
+    </AppContainer>
   );
 };
